修复广播记录时间基准不一致
Why: 设备生命周期时间使用 FreeRTOS tick，而 SNTP 格式化使用 esp_timer，两个时钟基准混用会使 recorded_at 与广播时间产生明显偏差，且 32 位毫秒值长期运行会回绕。

What: 将生命周期绝对时间戳升级为 64 位 esp_timer 毫秒值，并新增时间基准计算核心及主机测试。

How: 保留 tick 时钟用于扫描超时和 event_uptime_s；使用同一 esp_timer 基准生成和格式化广播绝对时间；执行 ./tests/run_host_tests.sh 与 idf.py build 验证。

+++ main/device/device_lifecycle.h
@@ -20,7 +20,7 @@ typedef struct {
     uint32_t broadcast_started_ms;
     uint32_t last_seen_ms;
     uint32_t end_detected_ms;
-    uint32_t broadcast_started_wall_ms;
-    uint32_t last_seen_wall_ms;
-    uint32_t end_detected_wall_ms;
+    uint64_t broadcast_started_wall_ms;
+    uint64_t last_seen_wall_ms;
+    uint64_t end_detected_wall_ms;
 } device_lifecycle_event_t;
